Fix getDERfromPEM
I get a wrong hash in Ubuntu 22.04. I suspect that the regular expression of awk in `getDERfromPEM` caused this problem.
Without the [\r|\n]*, awk will never set i=1. As a result, the awk will output an empty string, the downstream functions will process an empty string and finally output a wrong hash (of course the hash of an empty string is not the hash of the certificate).